moved hookRunner into /hooks folder
